@@ -13,7 +13,7 @@ task('javadocJar', type: Jar, dependsOn: javadoc) {
 tasks.withType(Javadoc).configureEach {
     javadocTool = javaToolchains.javadocToolFor {
         // Use latest JDK for javadoc generation
-        languageVersion = JavaLanguageVersion.of(17)
+        languageVersion = JavaLanguageVersion.of(21)
     }
 }
 
